+++ docs/changelog.rst
@@ -4,6 +4,9 @@ Changelog
 `Unreleased`_ - TBD
 -------------------
 
+`0.7.0`_ - 2022-04-26
+---------------------
+
 **Added**
 
 - Support for custom query printers. `#21`_
@@ -136,7 +139,8 @@ Invalid queries:
 
 - Initial public release
 
-.. _Unreleased: https://github.com/stranger6667/hypothesis-graphql/compare/v0.6.1...HEAD
+.. _Unreleased: https://github.com/stranger6667/hypothesis-graphql/compare/v0.7.0...HEAD
+.. _0.7.0: https://github.com/stranger6667/hypothesis-graphql/compare/v0.6.1...v0.7.0
 .. _0.6.1: https://github.com/stranger6667/hypothesis-graphql/compare/v0.6.0...v0.6.1
 .. _0.6.0: https://github.com/stranger6667/hypothesis-graphql/compare/v0.5.1...v0.6.0
 .. _0.5.1: https://github.com/stranger6667/hypothesis-graphql/compare/v0.5.0...v0.5.1